backend/requirements.txt
docker
fastapi
uvicorn[standard]
pydantic
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
# email-validator is commented out, can be added if needed for Pydantic's EmailStr strict validation
